requirements.txt
# Core AI & LangChain
langchain==0.1.0  # replace with latest stable
openai==1.8.5      # make sure your API keys work
faiss-cpu==1.7.4   # for vector DB embeddings

# Document loaders
unstructured==0.7.9
langchain-community==0.0.175

# Web scraping (if you want)
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Backend
fastapi==0.103.2
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.5.1

# Frontend
streamlit==1.29.0
streamlit-option-menu==0.5.4

# Map / visualizations (optional)
folium==0.14.0
pandas==2.1.1
streamlit
requests
wikipedia